+++ src/com/sanduhr/main/commands/ownerwhitelist/Mentioneveryone.java
@@ -1,12 +1,14 @@
 package com.sanduhr.main.commands.ownerwhitelist;
 
 import com.sanduhr.main.Lib;
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.ChannelType;
 import net.dv8tion.jda.core.events.ReadyEvent;
 import net.dv8tion.jda.core.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.core.events.message.MessageUpdateEvent;
 import net.dv8tion.jda.core.hooks.ListenerAdapter;
 
+@SuppressWarnings("ALL")
 public class Mentioneveryone extends ListenerAdapter {
 
     @Override
@@ -36,7 +38,11 @@ public class Mentioneveryone extends ListenerAdapter {
         }
 
         Lib.receivedcmd++;
-        e.getMessage().delete().queue();
+
+        if (e.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) {
+            e.getMessage().delete().queue();
+        }
+
         e.getChannel().sendMessage(e.getGuild().getPublicRole().getName()).queue();
         Lib.executedcmd++;
     }
@@ -49,20 +55,21 @@ public class Mentioneveryone extends ListenerAdapter {
         initter();
     }
 
-    public void initter() {
+    private void initter() {
         Lib.getCmdMap().put(getName(), getDescription());
         Lib.getSynMap().put(getName(), getSyntax());
     }
 
-    public String getName() {
+    private String getName() {
         return Mentioneveryone.class.getSimpleName().toLowerCase();
     }
 
-    public String getDescription() {
+    @SuppressWarnings("SameReturnValue")
+    private String getDescription() {
         return "mssingno";
     }
 
-    public String getSyntax() {
+    private String getSyntax() {
         return "`" + Lib.PREFIX + getName() + "`";
     }
 }
